requirements.txt
openai
pydantic>2
requests
python-dotenv
chromadb
numpy